change return type in GenreRepositoryJPA

+++ hw06-spring_ORM_JPA/src/main/java/ru/otus/repository/GenreRepositoryJPA.java
@@ -15,5 +15,5 @@ public interface GenreRepositoryJPA {
     Genre findById(long id);
     Genre findByName(String name);
     List<Genre> findAll();
-    int count();
+    Long count();
 }
